feat : add getZoom in mapController (#141) * change return type of getZoom from int to double

+++ lib/src/interface_osm/osm_interface.dart
@@ -159,7 +159,7 @@ abstract class OSMPlatform extends PlatformInterface {
     String id,
   );
 
-  Future<int> getZoom(
+  Future<double> getZoom(
     int idOSM,
   );
